Add ExampleActivity and register a listener to SystemMessenger
